Update README.md
Bird Box is an American horror-thriller film directed by Susanne Bier starring Sandra Bullock in the lead character as Malorie Hayes. The film was set in a post-apocalyptic situation where people need to be blindfolded to survive the extraterrestrial attack. Yes! Just by seeing the creatures, people get into hipnotic mode and die immediately by committing suicide. The story was adapted from the novel written by Josh Malerman which was well-conceived and conceptualized for a perfect seat edge thriller. The best part is we will never witness the strange creature on the screen which kills so many people. All we can witness is some black shadow with a strong wind. Nevertheless, it is just enough for the director to keep the audience engaged for a 90 mins captivating thriller.